Rename teamEditor files to align with app naming convention
