+++ template-engine/public/src/teamEditor/queryController.js
