+++ template-engine/views/teamEditor.handlebars
@@ -57,4 +57,4 @@
     <input type="hidden" id="team-id" value="{{data.id}}">
 </div>
 {{> confirmationModal }}
-<script type="module" src="src/teamEditor/teamEditor.js"></script>
+<script type="module" src="src/teamEditor/eventHandler.js"></script>
